+++ common/db/behavioral/operations_strategies/controller/include/strategy_control.hpp
@@ -1,11 +1,6 @@
 #pragma once
 
-#include <iostream>
-#include <exception>
-#include <functional>
-#include <vector>
-#include <memory>
-#include <queue>
+
 #include "batcher_strategy.hpp"
 #include "logger_strategy.hpp"
 #include "db_interface.hpp"
