+++ common/hash_creator/interface/hash_creator_interface.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <string>
-#include <openssl/sha.h>
+
 namespace drug_lib::common::hash_creator
 {
     class HashCreator
@@ -9,10 +9,10 @@ namespace drug_lib::common::hash_creator
     protected:
         /// @brief secret key для hmac
         std::string key = "af2bdbe1aa9b6ec1e2ade1d694f41fc71a831d0268e9891562113d8a62add1bf";
+
     public:
         virtual ~HashCreator() = default;
-        virtual std::string hash_function(const std::string &data, const std::string &salt) = 0;
-        HashCreator()= default;
+        virtual std::string hash_function(const std::string& data, const std::string& salt) = 0;
+        HashCreator() = default;
     };
-
 }
