+++ data/objects/base/include/medicament_base.hpp
@@ -1,8 +1,7 @@
 #pragma once
 
 #include "common_obj.hpp"
-#include "data_property.hpp"
-
+#include "property_factory.hpp"
 
 namespace drug_lib::data::objects
 {
@@ -60,13 +59,17 @@ namespace drug_lib::data::objects
                 {
                     id_ = field->as<int32_t>();
                 }
-                if (field_name == "name")
+                else if (field_name == "name")
                 {
                     name_ = field->as<std::string>();
                 }
-                if (field_name == properties::properties)
+                else if (field_name == properties::properties)
                 {
                     create_collection(field);
+                }
+                else
+                {
+                    throw std::invalid_argument("Unknown field name: " + field_name);
                 }
             }
         }
