+++ data/objects/base/include/objects.hpp
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "medicament_base.hpp"
+#include "patients_base.hpp"
+#include "organization_base.hpp"
+
