+++ data/objects/base/source/objects.cpp
@@ -1,0 +1,2 @@
+
+#include "objects.hpp"
